python API cleanups
