+++ wrappers/python/aries_askar/key.py
@@ -121,45 +121,3 @@ class Key:
             f"<Key(handle={self._handle}, alg={self.algorithm}, "
             f"ephemeral={self.ephemeral})>"
         )
-
-
-def crypto_box_random_nonce() -> bytes:
-    return bytes(bindings.key_crypto_box_random_nonce())
-
-
-def crypto_box(
-    recip_key: Key,
-    sender_key: Key,
-    message: Union[bytes, str],
-    nonce: bytes,
-) -> bytes:
-    return bytes(
-        bindings.key_crypto_box(recip_key._handle, sender_key._handle, message, nonce)
-    )
-
-
-def crypto_box_open(
-    recip_key: Key,
-    sender_key: Key,
-    message: Union[bytes, str],
-    nonce: bytes,
-) -> bytes:
-    return bytes(
-        bindings.key_crypto_box_open(
-            recip_key._handle, sender_key._handle, message, nonce
-        )
-    )
-
-
-def crypto_box_seal(
-    recip_key: Key,
-    message: Union[bytes, str],
-) -> bytes:
-    return bytes(bindings.key_crypto_box_seal(recip_key._handle, message))
-
-
-def crypto_box_seal_open(
-    recip_key: Key,
-    ciphertext: bytes,
-) -> bytes:
-    return bytes(bindings.key_crypto_box_seal_open(recip_key._handle, ciphertext))
