+++ wrappers/python/aries_askar/store.py
@@ -268,6 +268,11 @@ class Store:
         self._opener: OpenSession = None
         self._uri = uri
 
+    @classmethod
+    def generate_raw_key(cls, seed: Union[str, bytes] = None) -> str:
+        """Generate a new raw key for a Store."""
+        bindings.generate_raw_key(seed)
+
     @property
     def handle(self) -> StoreHandle:
         """Accessor for the store handle."""
